Factored out ExampleResourceLocator
